Move site active positions logic into SitePositionsController
This will be used later in `HostPortalBinding` to show positions for
every participant also for host portals.
