+++ lib/editor-binding.js
@@ -9,14 +9,11 @@ function doNothing () {}
 
 module.exports =
 class EditorBinding {
-  constructor ({editor, portal, isHost, didResize, didScroll, didDispose}) {
+  constructor ({editor, portal, isHost}) {
     this.editor = editor
     this.portal = portal
     this.isHost = isHost
     this.emitter = new Emitter()
-    this.emitDidResize = didResize || doNothing
-    this.emitDidScroll = didScroll || doNothing
-    this.emitDidDispose = didDispose || doNothing
     this.selectionsMarkerLayer = this.editor.selectionsMarkerLayer.bufferMarkerLayer
     this.markerLayersBySiteId = new Map()
     this.markersByLayerAndId = new WeakMap()
@@ -33,7 +30,8 @@ class EditorBinding {
     if (!this.isHost) this.restoreOriginalEditorMethods(this.editor)
     if (this.localCursorLayerDecoration) this.localCursorLayerDecoration.destroy()
 
-    this.emitDidDispose()
+    this.emitter.emit('did-dispose')
+    this.emitter.dispose()
   }
 
   setEditorProxy (editorProxy) {
@@ -130,20 +128,32 @@ class EditorBinding {
     const {element} = this.editor
     await element.component.getNextUpdatePromise()
     this.editorProxy.didScroll()
-    this.emitDidScroll()
+    this.emitter.emit('did-scroll')
   }
 
   async editorDidChangeScrollLeft () {
     const {element} = this.editor
     await element.component.getNextUpdatePromise()
     this.editorProxy.didScroll()
-    this.emitDidScroll()
+    this.emitter.emit('did-scroll')
   }
 
   async editorDidResize () {
     const {element} = this.editor
     await element.component.getNextUpdatePromise()
-    this.emitDidResize()
+    this.emitter.emit('did-resize')
+  }
+
+  onDidDispose (callback) {
+    return this.emitter.on('did-dispose', callback)
+  }
+
+  onDidScroll (callback) {
+    return this.emitter.on('did-scroll', callback)
+  }
+
+  onDidResize (callback) {
+    return this.emitter.on('did-resize', callback)
   }
 
   updateSelectionsForSiteId (siteId, selections) {
