+++ lib/guest-portal-binding.js
@@ -3,7 +3,7 @@ const {Errors, FollowState} = require('@atom/teletype-client')
 const BufferBinding = require('./buffer-binding')
 const EditorBinding = require('./editor-binding')
 const EmptyPortalPaneItem = require('./empty-portal-pane-item')
-const SitePositionsComponent = require('./site-positions-component')
+const SitePositionsController = require('./site-positions-controller')
 
 module.exports =
 class GuestPortalBinding {
@@ -20,7 +20,6 @@ class GuestPortalBinding {
     this.emitter = new Emitter()
     this.subscriptions = new CompositeDisposable()
     this.lastEditorProxyChangePromise = Promise.resolve()
-    this.positionsBySiteId = {}
   }
 
   async initialize () {
@@ -28,10 +27,12 @@ class GuestPortalBinding {
       this.portal = await this.client.joinPortal(this.portalId)
       if (!this.portal) return false
 
-      this.aboveViewportSitePositionsComponent = this.buildSitePositionsComponent('upper-right')
-      this.insideViewportSitePositionsComponent = this.buildSitePositionsComponent('middle-right')
-      this.outsideViewportSitePositionsComponent = this.buildSitePositionsComponent('lower-right')
-      this.showSitePositions()
+      this.sitePositionsController = new SitePositionsController({
+        portal: this.portal,
+        workspace: this.workspace,
+        editorBindingForEditorProxy: (editorProxy) => this.editorBindingsByEditorProxy.get(editorProxy)
+      })
+      this.sitePositionsController.show()
 
       await this.portal.setDelegate(this)
       await this.toggleEmptyPortalPaneItem()
@@ -46,9 +47,7 @@ class GuestPortalBinding {
 
   dispose () {
     this.subscriptions.dispose()
-    this.aboveViewportSitePositionsComponent.destroy()
-    this.insideViewportSitePositionsComponent.destroy()
-    this.outsideViewportSitePositionsComponent.destroy()
+    this.sitePositionsController.destroy()
     if (this.emptyPortalItem) this.emptyPortalItem.destroy()
 
     this.emitDidDispose()
@@ -86,40 +85,7 @@ class GuestPortalBinding {
   }
 
   updateActivePositions (positionsBySiteId) {
-    const aboveViewportSiteIds = []
-    const insideViewportSiteIds = []
-    const outsideViewportSiteIds = []
-
-    for (let siteId in positionsBySiteId) {
-      siteId = parseInt(siteId)
-      if (siteId === this.portal.siteId) continue
-
-      const {editorProxy, position} = positionsBySiteId[siteId]
-      const editorBinding = this.editorBindingsByEditorProxy.get(editorProxy)
-      if (position && editorBinding && editorBinding.editor === this.getActivePaneItem()) {
-        switch (editorBinding.getDirectionFromViewportToPosition(position)) {
-          case 'upward':
-            aboveViewportSiteIds.push(siteId)
-            break
-          case 'inside':
-            insideViewportSiteIds.push(siteId)
-            break
-          case 'downward':
-          case 'leftward':
-          case 'rightward':
-            outsideViewportSiteIds.push(siteId)
-            break
-        }
-      } else {
-        outsideViewportSiteIds.push(siteId)
-      }
-    }
-
-    const followedSiteId = this.portal.getFollowedSiteId()
-    this.aboveViewportSitePositionsComponent.update({siteIds: aboveViewportSiteIds, followedSiteId})
-    this.insideViewportSitePositionsComponent.update({siteIds: insideViewportSiteIds, followedSiteId})
-    this.outsideViewportSitePositionsComponent.update({siteIds: outsideViewportSiteIds, followedSiteId})
-    this.positionsBySiteId = positionsBySiteId
+    this.sitePositionsController.updateActivePositions(positionsBySiteId)
   }
 
   updateTether (followState, editorProxy, position) {
@@ -161,18 +127,17 @@ class GuestPortalBinding {
       editorBinding = new EditorBinding({
         editor,
         portal: this.portal,
-        isHost: false,
-        didScroll: () => this.updateActivePositions(this.positionsBySiteId),
-        didResize: () => this.updateActivePositions(this.positionsBySiteId),
-        didDispose: () => {
-          this.editorBindingsByEditorProxy.delete(editorProxy)
-          this.editorProxiesByEditor.delete(editor)
-        }
+        isHost: false
       })
       editorBinding.setEditorProxy(editorProxy)
       editorProxy.setDelegate(editorBinding)
+
       this.editorBindingsByEditorProxy.set(editorProxy, editorBinding)
       this.editorProxiesByEditor.set(editor, editorProxy)
+      editorBinding.onDidDispose(() => {
+        this.editorBindingsByEditorProxy.delete(editorProxy)
+        this.editorProxiesByEditor.delete(editor)
+      })
     }
     return editor
   }
@@ -269,9 +234,9 @@ class GuestPortalBinding {
     if (paneItem !== this.getEmptyPortalPaneItem()) {
       const editorProxy = this.editorProxiesByEditor.get(paneItem)
       if (editorProxy) {
-        this.showSitePositions()
+        this.sitePositionsController.show()
       } else {
-        this.hideSitePositions()
+        this.sitePositionsController.hide()
       }
 
       this.portal.activateEditorProxy(editorProxy)
@@ -300,36 +265,5 @@ class GuestPortalBinding {
 
   onDidChange (callback) {
     return this.emitter.on('did-change', callback)
-  }
-
-  // Private
-  showSitePositions () {
-    const workspaceElement = this.workspace.getElement()
-    workspaceElement.appendChild(this.aboveViewportSitePositionsComponent.element)
-    workspaceElement.appendChild(this.insideViewportSitePositionsComponent.element)
-    workspaceElement.appendChild(this.outsideViewportSitePositionsComponent.element)
-  }
-
-  // Private
-  hideSitePositions () {
-    this.aboveViewportSitePositionsComponent.element.remove()
-    this.insideViewportSitePositionsComponent.element.remove()
-    this.outsideViewportSitePositionsComponent.element.remove()
-  }
-
-  // Private
-  buildSitePositionsComponent (position) {
-    return new SitePositionsComponent({
-      position,
-      displayedParticipantsCount: 3,
-      portal: this.portal,
-      onSelectSiteId: (siteId) => {
-        if (siteId === this.portal.getFollowedSiteId()) {
-          this.portal.unfollow()
-        } else {
-          this.portal.follow(siteId)
-        }
-      }
-    })
   }
 }
