+++ lib/host-portal-binding.js
@@ -111,20 +111,17 @@ class HostPortalBinding {
     if (editorBinding) {
       return editorBinding.editorProxy
     } else {
-      editorBinding = new EditorBinding({editor, portal: this.portal, isHost: true})
       const bufferProxy = this.findOrCreateBufferProxyForBuffer(editor.getBuffer())
       const editorProxy = this.portal.createEditorProxy({bufferProxy})
-      editorBinding = new EditorBinding({
-        editor,
-        portal: this.portal,
-        isHost: true,
-        didDispose: () => this.editorBindingsByEditorProxy.delete(editorProxy)
-      })
+      editorBinding = new EditorBinding({editor, portal: this.portal, isHost: true})
       editorBinding.setEditorProxy(editorProxy)
       editorProxy.setDelegate(editorBinding)
 
       this.editorBindingsByEditor.set(editor, editorBinding)
       this.editorBindingsByEditorProxy.set(editorProxy, editorBinding)
+      editorBinding.onDidDispose(() => {
+        this.editorBindingsByEditorProxy.delete(editorProxy)
+      })
 
       return editorProxy
     }
